+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 IMPROVEMENTS:
 
+- Rds supports to create 10.0 PostgreSQL instance. ([#230](https://github.com/terraform-providers/terraform-provider-alicloud/pull/230))
 - *New Resource*: _alicloud_fc_trigger_ ([#228](https://github.com/terraform-providers/terraform-provider-alicloud/pull/228))
 - *New Resource*: _alicloud_fc_function_ ([#227](https://github.com/terraform-providers/terraform-provider-alicloud/pull/227))
-- *New Resource*: _alicloud_fc_service_ ([#226](https://github.com/terraform-providers/terraform-provider-alicloud/pull/226))
+- *New Resource*: _alicloud_fc_service_ 30([#226](https://github.com/terraform-providers/terraform-provider-alicloud/pull/226))
 - Support new field 'instance_name' for _alicloud_ots_table_ ([#225](https://github.com/terraform-providers/terraform-provider-alicloud/pull/225))
 - *New Resource*: _alicloud_ots_instance_attachment_ ([#224](https://github.com/terraform-providers/terraform-provider-alicloud/pull/224))
 - *New Resource*: _alicloud_ots_instance_ ([#223](https://github.com/terraform-providers/terraform-provider-alicloud/pull/223))