+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.9.7 (Unreleased)
+## 1.10.0 (Unreleased)
 
 IMPROVEMENTS: